Move methods _search_indices_curvilinear, _search_indices_rectilinear from field.py to _index_search.py
