+++ parcels/field.py
@@ -12,7 +12,6 @@ import xarray as xr
 
 import parcels.tools.interpolation_utils as i_u
 from parcels._compat import add_note
-from parcels._index_search import search_indices_vertical_s, search_indices_vertical_z
 from parcels._interpolation import (
     InterpolationContext2D,
     InterpolationContext3D,
@@ -41,11 +40,10 @@ from parcels.tools.statuscodes import (
     FieldSamplingError,
     TimeExtrapolationError,
     _raise_field_out_of_bound_error,
-    _raise_field_out_of_bound_surface_error,
-    _raise_field_sampling_error,
 )
 from parcels.tools.warnings import FieldSetWarning, _deprecated_param_netcdf_decodewarning
 
+from ._index_search import _search_indices_curvilinear, _search_indices_rectilinear
 from .fieldfilebuffer import (
     DaskFileBuffer,
     DeferredDaskFileBuffer,
@@ -1006,193 +1004,9 @@ class Field:
     def search_indices_rectilinear(self, *_):
         raise NotImplementedError
 
-    def _search_indices_rectilinear(
-        self, time: float, z: float, y: float, x: float, ti=-1, particle=None, search2D=False
-    ):
-        grid = self.grid
-
-        if grid.xdim > 1 and (not grid.zonal_periodic):
-            if x < grid.lonlat_minmax[0] or x > grid.lonlat_minmax[1]:
-                _raise_field_out_of_bound_error(z, y, x)
-        if grid.ydim > 1 and (y < grid.lonlat_minmax[2] or y > grid.lonlat_minmax[3]):
-            _raise_field_out_of_bound_error(z, y, x)
-
-        if grid.xdim > 1:
-            if grid.mesh != "spherical":
-                lon_index = grid.lon < x
-                if lon_index.all():
-                    xi = len(grid.lon) - 2
-                else:
-                    xi = lon_index.argmin() - 1 if lon_index.any() else 0
-                xsi = (x - grid.lon[xi]) / (grid.lon[xi + 1] - grid.lon[xi])
-                if xsi < 0:
-                    xi -= 1
-                    xsi = (x - grid.lon[xi]) / (grid.lon[xi + 1] - grid.lon[xi])
-                elif xsi > 1:
-                    xi += 1
-                    xsi = (x - grid.lon[xi]) / (grid.lon[xi + 1] - grid.lon[xi])
-            else:
-                lon_fixed = grid.lon.copy()
-                indices = lon_fixed >= lon_fixed[0]
-                if not indices.all():
-                    lon_fixed[indices.argmin() :] += 360
-                if x < lon_fixed[0]:
-                    lon_fixed -= 360
-
-                lon_index = lon_fixed < x
-                if lon_index.all():
-                    xi = len(lon_fixed) - 2
-                else:
-                    xi = lon_index.argmin() - 1 if lon_index.any() else 0
-                xsi = (x - lon_fixed[xi]) / (lon_fixed[xi + 1] - lon_fixed[xi])
-                if xsi < 0:
-                    xi -= 1
-                    xsi = (x - lon_fixed[xi]) / (lon_fixed[xi + 1] - lon_fixed[xi])
-                elif xsi > 1:
-                    xi += 1
-                    xsi = (x - lon_fixed[xi]) / (lon_fixed[xi + 1] - lon_fixed[xi])
-        else:
-            xi, xsi = -1, 0
-
-        if grid.ydim > 1:
-            lat_index = grid.lat < y
-            if lat_index.all():
-                yi = len(grid.lat) - 2
-            else:
-                yi = lat_index.argmin() - 1 if lat_index.any() else 0
-
-            eta = (y - grid.lat[yi]) / (grid.lat[yi + 1] - grid.lat[yi])
-            if eta < 0:
-                yi -= 1
-                eta = (y - grid.lat[yi]) / (grid.lat[yi + 1] - grid.lat[yi])
-            elif eta > 1:
-                yi += 1
-                eta = (y - grid.lat[yi]) / (grid.lat[yi + 1] - grid.lat[yi])
-        else:
-            yi, eta = -1, 0
-
-        if grid.zdim > 1 and not search2D:
-            if grid._gtype == GridType.RectilinearZGrid:
-                # Never passes here, because in this case, we work with scipy
-                try:
-                    (zi, zeta) = search_indices_vertical_z(self.grid, self.gridindexingtype, z)
-                except FieldOutOfBoundError:
-                    _raise_field_out_of_bound_error(z, y, x)
-                except FieldOutOfBoundSurfaceError:
-                    _raise_field_out_of_bound_surface_error(z, y, x)
-            elif grid._gtype == GridType.RectilinearSGrid:
-                (zi, zeta) = search_indices_vertical_s(
-                    self.grid, self.interp_method, time, z, y, x, ti, yi, xi, eta, xsi
-                )
-        else:
-            zi, zeta = -1, 0
-
-        if not ((0 <= xsi <= 1) and (0 <= eta <= 1) and (0 <= zeta <= 1)):
-            _raise_field_sampling_error(z, y, x)
-
-        if particle:
-            particle.xi[self.igrid] = xi
-            particle.yi[self.igrid] = yi
-            particle.zi[self.igrid] = zi
-
-        return (zeta, eta, xsi, zi, yi, xi)
-
     @deprecated_made_private  # TODO: Remove 6 months after v3.1.0
     def search_indices_curvilinear(self, *_):
         raise NotImplementedError
-
-    def _search_indices_curvilinear(self, time, z, y, x, ti=-1, particle=None, search2D=False):
-        if particle:
-            xi = particle.xi[self.igrid]
-            yi = particle.yi[self.igrid]
-        else:
-            xi = int(self.grid.xdim / 2) - 1
-            yi = int(self.grid.ydim / 2) - 1
-        xsi = eta = -1
-        grid = self.grid
-        invA = np.array([[1, 0, 0, 0], [-1, 1, 0, 0], [-1, 0, 0, 1], [1, -1, 1, -1]])
-        maxIterSearch = 1e6
-        it = 0
-        tol = 1.0e-10
-        if not grid.zonal_periodic:
-            if x < grid.lonlat_minmax[0] or x > grid.lonlat_minmax[1]:
-                if grid.lon[0, 0] < grid.lon[0, -1]:
-                    _raise_field_out_of_bound_error(z, y, x)
-                elif x < grid.lon[0, 0] and x > grid.lon[0, -1]:  # This prevents from crashing in [160, -160]
-                    _raise_field_out_of_bound_error(z, y, x)
-        if y < grid.lonlat_minmax[2] or y > grid.lonlat_minmax[3]:
-            _raise_field_out_of_bound_error(z, y, x)
-
-        while xsi < -tol or xsi > 1 + tol or eta < -tol or eta > 1 + tol:
-            px = np.array([grid.lon[yi, xi], grid.lon[yi, xi + 1], grid.lon[yi + 1, xi + 1], grid.lon[yi + 1, xi]])
-            if grid.mesh == "spherical":
-                px[0] = px[0] + 360 if px[0] < x - 225 else px[0]
-                px[0] = px[0] - 360 if px[0] > x + 225 else px[0]
-                px[1:] = np.where(px[1:] - px[0] > 180, px[1:] - 360, px[1:])
-                px[1:] = np.where(-px[1:] + px[0] > 180, px[1:] + 360, px[1:])
-            py = np.array([grid.lat[yi, xi], grid.lat[yi, xi + 1], grid.lat[yi + 1, xi + 1], grid.lat[yi + 1, xi]])
-            a = np.dot(invA, px)
-            b = np.dot(invA, py)
-
-            aa = a[3] * b[2] - a[2] * b[3]
-            bb = a[3] * b[0] - a[0] * b[3] + a[1] * b[2] - a[2] * b[1] + x * b[3] - y * a[3]
-            cc = a[1] * b[0] - a[0] * b[1] + x * b[1] - y * a[1]
-            if abs(aa) < 1e-12:  # Rectilinear cell, or quasi
-                eta = -cc / bb
-            else:
-                det2 = bb * bb - 4 * aa * cc
-                if det2 > 0:  # so, if det is nan we keep the xsi, eta from previous iter
-                    det = np.sqrt(det2)
-                    eta = (-bb + det) / (2 * aa)
-            if abs(a[1] + a[3] * eta) < 1e-12:  # this happens when recti cell rotated of 90deg
-                xsi = ((y - py[0]) / (py[1] - py[0]) + (y - py[3]) / (py[2] - py[3])) * 0.5
-            else:
-                xsi = (x - a[0] - a[2] * eta) / (a[1] + a[3] * eta)
-            if xsi < 0 and eta < 0 and xi == 0 and yi == 0:
-                _raise_field_out_of_bound_error(0, y, x)
-            if xsi > 1 and eta > 1 and xi == grid.xdim - 1 and yi == grid.ydim - 1:
-                _raise_field_out_of_bound_error(0, y, x)
-            if xsi < -tol:
-                xi -= 1
-            elif xsi > 1 + tol:
-                xi += 1
-            if eta < -tol:
-                yi -= 1
-            elif eta > 1 + tol:
-                yi += 1
-            (yi, xi) = self._reconnect_bnd_indices(yi, xi, grid.ydim, grid.xdim, grid.mesh)
-            it += 1
-            if it > maxIterSearch:
-                print(f"Correct cell not found after {maxIterSearch} iterations")
-                _raise_field_out_of_bound_error(0, y, x)
-        xsi = max(0.0, xsi)
-        eta = max(0.0, eta)
-        xsi = min(1.0, xsi)
-        eta = min(1.0, eta)
-
-        if grid.zdim > 1 and not search2D:
-            if grid._gtype == GridType.CurvilinearZGrid:
-                try:
-                    (zi, zeta) = search_indices_vertical_z(self.grid, self.gridindexingtype, z)
-                except FieldOutOfBoundError:
-                    _raise_field_out_of_bound_error(z, y, x)
-            elif grid._gtype == GridType.CurvilinearSGrid:
-                (zi, zeta) = search_indices_vertical_s(
-                    self.grid, self.interp_method, time, z, y, x, ti, yi, xi, eta, xsi
-                )
-        else:
-            zi = -1
-            zeta = 0
-
-        if not ((0 <= xsi <= 1) and (0 <= eta <= 1) and (0 <= zeta <= 1)):
-            _raise_field_sampling_error(z, y, x)
-
-        if particle:
-            particle.xi[self.igrid] = xi
-            particle.yi[self.igrid] = yi
-            particle.zi[self.igrid] = zi
-
-        return (zeta, eta, xsi, zi, yi, xi)
 
     @deprecated_made_private  # TODO: Remove 6 months after v3.1.0
     def search_indices(self, *_):
@@ -1200,9 +1014,9 @@ class Field:
 
     def _search_indices(self, time, z, y, x, ti=-1, particle=None, search2D=False):
         if self.grid._gtype in [GridType.RectilinearSGrid, GridType.RectilinearZGrid]:
-            return self._search_indices_rectilinear(time, z, y, x, ti, particle=particle, search2D=search2D)
+            return _search_indices_rectilinear(self, time, z, y, x, ti, particle=particle, search2D=search2D)
         else:
-            return self._search_indices_curvilinear(time, z, y, x, ti, particle=particle, search2D=search2D)
+            return _search_indices_curvilinear(self, time, z, y, x, ti, particle=particle, search2D=search2D)
 
     @deprecated_made_private  # TODO: Remove 6 months after v3.1.0
     def interpolator2D(self, *_):
